requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==22.1.0
blinker==1.7.0
click==8.1.7
Django==4.0.1
Flask==3.0.2
Flask-Cors==4.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==0.4.3
packaging==23.2
pathspec==0.9.0
pillow==10.2.0
platformdirs==2.5.1
pycodestyle==2.8.0
pytesseract==0.3.10
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==3.0.1
zipp==3.17.0
